add answer questions and error controller

+++ src/main/java/dqx/community/controller/PublishController.java
@@ -6,6 +6,7 @@ import dqx.community.mapper.UserMapper;
 import dqx.community.model.Question;
 import dqx.community.model.User;
 import dqx.community.service.QuestionService;
+import org.apache.ibatis.annotations.Lang;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -23,7 +24,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,Model model){
+    public String edit(@PathVariable(name = "id") Long id, Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle() );
         model.addAttribute("description", question.getDescription());
@@ -42,7 +43,7 @@ public class PublishController {
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "id",required = false) Integer id,
+            @RequestParam(value = "id",required = false) Long id,
             HttpServletRequest request,
             Model model){
